１ステージを簡単なチュートリアル用ステージに置き換えた

+++ 再出発　遅延/Src/Application/GameObject/UI/Button/KeyEnter/KeyEneter.cpp
@@ -17,8 +17,8 @@ void KeyEnter::DrawSprite()
 {
 	if (m_drawFlg)
 	{
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, (int)m_pos.x, (int)m_pos.y,
-			(int)m_texSize.x * (int)m_scale.x, (int)m_texSize.y * (int)m_scale.y, nullptr, &m_color);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, m_pos.x, m_pos.y,
+			m_texSize.x * m_scale.x, m_texSize.y * m_scale.y, nullptr, &m_color);
 	}
 }
