+++ 再出発　遅延/Src/Application/GameObject/UI/ClearTime/ClearTime.cpp
@@ -57,25 +57,25 @@ void ClearTime::DrawSprite()
 	if (m_drawFlg)
 	{
 		//時間
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, (int)m_time.secPosR.x, (int)m_time.secPosR.y,
-			(int)m_time.texSize.x * (float)m_time.scale.x, (int)m_time.texSize.y * (float)m_time.scale.y, &m_time.secRectR, &m_time.color);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, m_time.secPosR.x, m_time.secPosR.y,
+			m_time.texSize.x * m_time.scale.x, m_time.texSize.y * m_time.scale.y, &m_time.secRectR, &m_time.color);
 
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, (int)m_time.secPosL.x, (int)m_time.secPosL.y,
-			(int)m_time.texSize.x * (float)m_time.scale.x, (int)m_time.texSize.y * (float)m_time.scale.y, &m_time.secRectL, &m_time.color);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, m_time.secPosL.x, m_time.secPosL.y,
+			m_time.texSize.x * m_time.scale.x, m_time.texSize.y * m_time.scale.y, &m_time.secRectL, &m_time.color);
 
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, (int)m_time.minPosR.x, (int)m_time.minPosR.y,
-			(int)m_time.texSize.x * (float)m_time.scale.x, (int)m_time.texSize.y * (float)m_time.scale.y, &m_time.minRectR, &m_time.color);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex,m_time.minPosR.x, m_time.minPosR.y,
+			m_time.texSize.x * m_time.scale.x, m_time.texSize.y * m_time.scale.y, &m_time.minRectR, &m_time.color);
 
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, (int)m_time.minPosL.x, (int)m_time.minPosL.y,
-			(int)m_time.texSize.x * (float)m_time.scale.x, (int)m_time.texSize.y * (float)m_time.scale.y, &m_time.minRectL, &m_time.color);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, m_time.minPosL.x, m_time.minPosL.y,
+			m_time.texSize.x * m_time.scale.x, m_time.texSize.y * m_time.scale.y, &m_time.minRectL, &m_time.color);
 
 		//コロン
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_colonTex, (int)m_colonPos.x, (int)m_colonPos.y,
-			(int)m_colonTexSize.x * (float)m_colonScale.x, (int)m_colonTexSize.y * (float)m_colonScale.y, &m_colonRect, &m_colonColor);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_colonTex, m_colonPos.x, m_colonPos.y,
+			m_colonTexSize.x * m_colonScale.x, m_colonTexSize.y * m_colonScale.y, &m_colonRect, &m_colonColor);
 
 		//スター
-		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_starTex, (int)m_starPos.x, (int)m_starPos.y,
-			(int)m_starTexSize.x * (float)m_starScale.x, (int)m_starTexSize.y * (float)m_starScale.y, nullptr, &m_starColor);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(&m_starTex, m_starPos.x, m_starPos.y,
+			m_starTexSize.x * m_starScale.x,m_starTexSize.y * m_starScale.y, nullptr, &m_starColor);
 	}
 }
